+++ lib/solidpod.dart
@@ -1,6 +1,6 @@
 /// Support for flutter apps accessing solid PODs.
 ///
-// Time-stamp: <Wednesday 2024-04-17 08:47:37 +1000 Graham Williams>
+// Time-stamp: <Monday 2024-04-22 15:19:23 +1000 Graham Williams>
 ///
 /// Copyright (C) 2024, Software Innovation Institute, ANU.
 ///
@@ -30,12 +30,20 @@
 
 library solidpod;
 
-export 'src/solid/login.dart' show SolidLogin;
+export 'src/solid/login.dart'
+    show
+        SolidLogin,
+        LoginButtonStyle,
+        ContinueButtonStyle,
+        RegisterButtonStyle,
+        InfoButtonStyle;
 export 'src/solid/popup_login.dart' show SolidPopupLogin;
 
 // TODO 20240417 gjw Can we please list or at least document what and why the
-// following are exported, PLEASE. ReadPod() I understand but waht from
+// following are exported, PLEASE. ReadPod() I understand, but what from
 // rest_api?
 
 export 'src/solid/api/rest_api.dart';
-export 'src/solid/read_pod.dart';
+export 'src/solid/utils.dart';
+export 'src/solid/read_pod.dart' show readPod;
+export 'src/solid/write_pod.dart' show writePod;